% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_diwf_boat_regs.R
\docType{data}
\name{me_diwf_boat_regs}
\alias{me_diwf_boat_regs}
\title{Annual State Boat Registrations (Maine Department of Inland Fisheries & Wildlife)}
\format{
A data frame with 19 rows and 2 variables:
\describe{
  \item{year}{DIWF boat registration year}
  \item{total_diwf_boat_regs}{Number of boats registered with DIWF}
}
}
\source{
<https://github.com/Social-Oceans-Lab/gomfish>
}
\usage{
me_diwf_boat_regs
}
\description{
A dataset with annual DIWF boat registration information in Maine from 2003-2021
}
\keyword{datasets}
