% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_garfo_boat_regs.R
\docType{data}
\name{me_garfo_boat_regs}
\alias{me_garfo_boat_regs}
\title{Annual Federal Boat Registrations (Greater Atlantic Regional Fisheries Office)}
\format{
A data frame with 19 rows and 2 variables:
\describe{
  \item{year}{GARFO boat registration year}
  \item{total_diwf_boat_regs}{Number of boats registered with GARFO}
}
}
\source{
<https://github.com/Social-Oceans-Lab/gomfish>
}
\usage{
me_garfo_boat_regs
}
\description{
A dataset with annual GARFO boat registration information in Maine from 2000-2021
}
\keyword{datasets}
