% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_housing.R
\docType{data}
\name{me_housing}
\alias{me_housing}
\title{Annual housing data in Maine}
\format{

}
\source{
<https://github.com/Social-Oceans-Lab/gomfish>
}
\usage{
me_housing
}
\description{
Annual, town level summary of several metrics related to housing affordability in Maine from 2000-2021. (Source: Maine State Housing Authority)
}
\details{
DETAILS
}
\keyword{datasets}
