% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_housing_affordability.R
\docType{data}
\name{me_housing_affordability}
\alias{me_housing_affordability}
\title{Maine: Annual Housing Affordability by Towns and Cities}
\format{

}
\source{
<https://github.com/Social-Oceans-Lab/gomfish>
}
\usage{
me_housing_affordability
}
\description{
A dataset with annual housing affordability metrics from Maine Housing from 2000-2022
}
\keyword{datasets}
