% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_lob_county.R
\docType{data}
\name{me_lob_county}
\alias{me_lob_county}
\title{Maine Fisheries Data: Annual Commercial Lobster Landings and Value by County}
\format{
A data frame with 443 rows and 4 variables:
\describe{
  \item{year}{Landing year.}
  \item{county}{Maine county.}
  \item{pounds}{Total pounds of lobster landed by the commercial fishery in a given year and county}
  \item{value}{Total value of lobster landed by the commercial fishery in a given year and county}
}
}
\source{
<https://github.com/Social-Oceans-Lab/gomfish>
}
\usage{
me_lob_county
}
\description{
A dataset with annual commercial lobster landings and value by Maine counties from 1964-present
}
\keyword{datasets}
