% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_marine_casualties.R
\docType{data}
\name{me_marine_casualties}
\alias{me_marine_casualties}
\title{Annual marine casualties in Maine}
\format{
A data frame with 47 rows and 3 variables:
\describe{
  \item{year}{Year}
  \item{region}{Coastal region in Maine}
  \item{num_casualties}{Total number of marine casualties}
}
}
\source{
<https://github.com/Social-Oceans-Lab/gomfish>
}
\usage{
me_marine_casualties
}
\description{
Annual, regional record of marine casualties that occurred off the coast of Maine on lobster vessels associated with a Maine lobster license from 1999-2015. This record includes all casualties that required US Coast Guard intervention including boat malfunctions, personal injury, and death.(Source: US Coast Guard)
}
\details{
DETAILS
}
\keyword{datasets}
