% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_monthly_vehicle_regs.R
\docType{data}
\name{me_monthly_vehicle_regs}
\alias{me_monthly_vehicle_regs}
\title{Monthly vehicle registrations in Maine}
\format{
A data frame with 46766 rows and 11 variables:
\describe{
  \item{year}{Year}
  \item{date}{YYYY-MM-01 date}
  \item{month}{Month}
  \item{region}{Coastal region in Maine}
  \item{county}{County in Maine}
  \item{town}{Town in Maine}
  \item{truck_total_regs}{Total pickup truck registrations}
  \item{truck_age_at_purchase}{Average age of pickup truck at time of purchase}
  \item{truck_base_price}{Average pickup truck base price}
  \item{truck_exise_amt}{Average pickup truck excise amount}
  \item{comm_truck_total_regs}{Total commercial truck registrations}
}
}
\source{
<https://github.com/Social-Oceans-Lab/gomfish>
}
\usage{
me_monthly_vehicle_regs
}
\description{
Monthly, town level dataset of registrations of pickup truck registrations and characteristics. (Source:Maine BMV)
}
\details{
DETAILS
}
\keyword{datasets}
