% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_real_estate.R
\docType{data}
\name{me_real_estate}
\alias{me_real_estate}
\title{Monthly real estate data in Maine}
\format{
A data frame with 25476 rows and 11 variables:
\describe{
  \item{year}{Year}
  \item{month}{Month}
  \item{date}{YYYY-MM-01 date}
  \item{region}{Coastal Region in Maine}
  \item{county}{County in Maine}
  \item{town}{Town in Maine}
  \item{avg_median_listing_price}{Median listing price of a home on the real estate market}
  \item{avg_median_days_on_market}{Median number of days a home is on the real estate market}
  \item{avg_median_listing_price_per_square_foot}{Median listing price per square foot of a home on the real estate market}
  \item{avg_listing_price}{Average listing price of a home on the real estate market}
  \item{total_listing_count}{Total number of home listings on the real estate market}
}
}
\source{
<https://github.com/Social-Oceans-Lab/gomfish>
}
\usage{
me_real_estate
}
\description{
This dataset presents a monthly, town level summary of real estate variables from July 2016-January 2022 (Source: Realtor.com)
}
\details{
DETAILS
}
\keyword{datasets}
