% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_resource_violations.R
\docType{data}
\name{me_resource_violations}
\alias{me_resource_violations}
\title{Monthly marine resource violations in Maine}
\format{
A data frame with 6436 rows and 6 variables:
\describe{
  \item{year}{Year}
  \item{month}{Month}
  \item{location}{Town in Maine}
  \item{lob_violations}{Total number of lobster related marine resource violations}
  \item{total_violations}{Total number of marine resource violations}
  \item{prop_lob_violations}{Proportion of total marine resource violations that are lobster related}
}
}
\source{
<https://github.com/Social-Oceans-Lab/gomfish>
}
\usage{
me_resource_violations
}
\description{
Monthly, town-level frequency of marine resource violations recorded by Maine Marine Patrol, including those that occurred on lobster fishing vessels from 2000-2021. (Source: Maine Marine Patrol)
}
\details{
DETAILS
}
\keyword{datasets}
