% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_risk_index.R
\docType{data}
\name{me_risk_index}
\alias{me_risk_index}
\title{Maine weather risk index}
\format{
A data frame with 30681 rows and 7 variables:
\describe{
  \item{year}{Year}
  \item{month}{Month}
  \item{week}{Week}
  \item{day}{Day}
  \item{date}{YYYY-MM-DD date}
  \item{zone}{Maine DMR Lobster Management Zone}
  \item{risk_index}{Composite risk index. See description for methods used to develop this index}
}
}
\source{
<https://github.com/Social-Oceans-Lab/gomfish>
}
\usage{
me_risk_index
}
\description{
A composite daily risk index, which describes the relative risk level of conditions at sea by zone and day from 2010-2021. The risk index was created by first calculating 12 hour daily averages of wind speed, wind gusts, air temperature, significant wave height, and wave period. Then, we conducted a principal component analysis to determine appropriate weighting for each weather component, resulting in a single value per zone and day. (Source: European Centre for Medium Weather Forecasts’s reanalysis product ERA5)
}
\details{
DETAILS
}
\keyword{datasets}
